+++ app/src/main/java/com/example/fe/bookclub_place/api/PlaceResponse.kt
@@ -1,7 +1,7 @@
 package com.example.fe.bookclub_place.api
 
 data class PlaceResponse(
-    val placeId: Int,        // 장소 ID
+    val placeId: Long,        // 장소 ID
     val title: String,       // 장소 이름
     val category: String,    // 카테고리 (카페, 서점 등)
     val address: String,     // 주소
