+++ app/src/main/java/com/example/fe/scrap/NewScrapDialogFragment.kt
@@ -30,7 +30,7 @@ import retrofit2.Response
 
 class NewScrapDialogFragment(
     private val bookId: Long?,
-    private val placeId: Int?,
+    private val placeId: Long?,
     private val onScrapCreated: () -> Unit // 북마크 상태 변경 콜백 추가
 ) : DialogFragment() {
